requirements.txt
transformers
torch
faiss-cpu
langchain
openai
pyyaml
pandas
numpy